Remove actions presenter and view
